backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic>=2.6
numpy>=1.24
pymupdf>=1.23
pillow>=10.0
pytesseract>=0.3
sentence-transformers>=2.6
openai>=1.0
tqdm>=4.0
